\name{IATscores-package}
\alias{IATscores-package}
\alias{IATscores}
\docType{package}
\title{
Compute Robust IAT scores
}
\description{
The function \code{\link{RobustScores}} computes variants of the robust IAT
scores according to four main parameters.
}
\details{
\tabular{ll}{
Package: \tab IATscores\cr
Type: \tab Package\cr
Version: \tab 0.2.3\cr
Date: \tab 2019-07-05\cr
License: \tab GPL-2 \cr
}
}