\name{IATscores-package}
\alias{IATscores-package}
\alias{IATscores}
\docType{package}
\title{
Compute Robust IAT scores
}
\description{
The function \code{\link{RobustScores}} computes variants of the robust IAT
scores according to four main parameters.
}
\details{
\tabular{ll}{
Package: \tab IATscores\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-31\cr
License: \tab What license is it under?\cr
}
}